fixed issues with multiple equal servers in group
